+++ apps/fluent-tester/src/E2E/BootPage/specs/BootTestPage.spec.uwp.ts
@@ -14,14 +14,14 @@ const PAGE_TIMEOUT = 3000;
 
 // Before testing begins, allow up to 60 seconds for bundle to load
 // (WebDriverIO)
-describe('Open the app', function() {
+describe('Open the app', function () {
   it('Boot app', () => {
     BootTestPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
     expect(BootTestPage.isPageLoaded()).toBeTruthy();
   });
 });
 
-describe('Click on each test page and check if it renders', function() {
+describe('Click on each test page and check if it renders', function () {
   it('Button Test Page', () => {
     BootTestPage.clickAndGoToButtonPage();
     ButtonTestPage.waitForPageDisplayed(PAGE_TIMEOUT);