+++ packages/theming/apple-theme/src/appleTheme.macos.ts
@@ -113,6 +113,27 @@ export const appleComponents = {
         },
       },
     },
+  Checkbox: {
+    checkbox: {
+      style: {
+        borderStyle: 'solid',
+        borderWidth: 0.5,
+        borderRadius: 3,
+        minHeight: 14,
+        minWidth: 14,
+        marginEnd: 5,
+      },
+    },
+    checkmarkIcon: {
+      width: 10,
+      height: 10,
+      style: {
+        marginVertical: 2,
+        marginHorizontal: 2,
+      },
+    },
+    // This disables other available states like: hovered, focused, pressed.
+    _precedence: ['disabled', 'boxAtEnd', 'checked'],
   },
 };