+++ packages/experimental/Button/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/experimental-button
 
-This log was last generated on Wed, 02 Jun 2021 17:30:36 GMT and should not be manually modified.
+This log was last generated on Fri, 18 Jun 2021 00:38:19 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.1
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to experimental, run (ruaraki@microsoft.com)
+
+## 0.6.0
+
+Tue, 15 Jun 2021 00:55:36 GMT
+
+### Minor changes
+
+- Handle ColorValue instead of string as color type (ruaraki@microsoft.com)
 
 ## 0.5.4