+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Tue, 25 May 2021 00:42:50 GMT and should not be manually modified.
+This log was last generated on Mon, 21 Jun 2021 11:19:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.2
+
+Mon, 21 Jun 2021 11:19:28 GMT
+
+### Patches
+
+- Update checkbox settings/theme for android (tamasane@gmail.com)
+
+## 0.7.1
+
+Fri, 18 Jun 2021 00:38:19 GMT
+
+### Patches
+
+- Apply prettier to components, run (ruaraki@microsoft.com)
+
+## 0.7.0
+
+Tue, 15 Jun 2021 00:55:36 GMT
+
+### Minor changes
+
+- Handle ColorValue input for color prop (ruaraki@microsoft.com)
 
 ## 0.6.0